fix(models): P6.1 monetary double→int audit (28 fields, 12 models)
VNĐ has no subunit — all prices/totals are whole numbers.
Changed 28 monetary/points fields from double to int across:
- MenuItem, CartItem, DeliveryOrder, OrderItem
- LoyaltyMember, PointTransaction, CheckinResult
- TierProgress, Promotion, Refund
- SupplierInvoice, SettlementBatch, GrnItem
- CartState (deliveryFee, discount, subtotal, total)

Kept double for rates/percentages (pointMultiplier, cashbackPercent,
vatRate, progressPercent) and quantities (stock, grn quantities).

Updated cart_screen _formatPrice(double→int), checkin_screen,
Formatters.currency(num→int), and all 16 test files.

flutter analyze: 0 errors (23 info)
flutter test: 183/183 passed

+++ apps/mobile/lib/features/cart/domain/cart_state.dart
@@ -18,8 +18,8 @@ class CartState {
   });
 
   final List<CartItem> items;
-  final double deliveryFee;
-  final double discount;
+  final int deliveryFee;
+  final int discount;
   final String paymentMethod;
   final String? deliveryAddress;
   final double? latitude;
@@ -30,9 +30,9 @@ class CartState {
   final bool isSubmitting;
   final String? orderError;
 
-  double get subtotal => items.fold(0, (sum, item) => sum + item.total);
+  int get subtotal => items.fold(0, (sum, item) => sum + item.total);
 
-  double get total => subtotal + deliveryFee - discount;
+  int get total => subtotal + deliveryFee - discount;
 
   int get itemCount => items.fold(0, (sum, item) => sum + item.quantity);
 
@@ -40,8 +40,8 @@ class CartState {
 
   CartState copyWith({
     List<CartItem>? items,
-    double? deliveryFee,
-    double? discount,
+    int? deliveryFee,
+    int? discount,
     String? paymentMethod,
     String? deliveryAddress,
     double? latitude,
